debug allocator

+++ src/core/json_serializer.cpp
@@ -208,7 +208,7 @@ void JsonSerializer::deserialize(float& value)
 void JsonSerializer::deserialize(int32_t& value)
 {
 	logErrorIfNot(!m_is_string_token);
-	logErrorIfNot(fromCString(m_token, strlen(m_token), &value));
+	logErrorIfNot(fromCString(m_token, (int)strlen(m_token), &value));
 	deserializeToken();
 }
 
@@ -233,7 +233,7 @@ void JsonSerializer::deserialize(const char* label, uint32_t& value)
 {
 	deserializeLabel(label);
 	logErrorIfNot(!m_is_string_token);
-	logErrorIfNot(fromCString(m_token, strlen(m_token), &value));
+	logErrorIfNot(fromCString(m_token, (int)strlen(m_token), &value));
 	deserializeToken();
 }
 
@@ -248,7 +248,7 @@ void JsonSerializer::deserialize(const char* label, int32_t& value)
 {
 	deserializeLabel(label);
 	logErrorIfNot(!m_is_string_token);
-	logErrorIfNot(fromCString(m_token, strlen(m_token), &value));
+	logErrorIfNot(fromCString(m_token, (int)strlen(m_token), &value));
 	deserializeToken();
 }
 
@@ -341,7 +341,7 @@ void JsonSerializer::deserializeArrayItem(uint32_t& value)
 {
 	deserializeArrayComma();
 	logErrorIfNot(!m_is_string_token);
-	logErrorIfNot(fromCString(m_token, strlen(m_token), &value));
+	logErrorIfNot(fromCString(m_token, (int)strlen(m_token), &value));
 	deserializeToken();
 }
 
@@ -350,7 +350,7 @@ void JsonSerializer::deserializeArrayItem(int32_t& value)
 {
 	deserializeArrayComma();
 	logErrorIfNot(!m_is_string_token);
-	logErrorIfNot(fromCString(m_token, strlen(m_token), &value));
+	logErrorIfNot(fromCString(m_token, (int)strlen(m_token), &value));
 	deserializeToken();
 }
 
@@ -359,7 +359,7 @@ void JsonSerializer::deserializeArrayItem(int64_t& value)
 {
 	deserializeArrayComma();
 	logErrorIfNot(!m_is_string_token);
-	logErrorIfNot(fromCString(m_token, strlen(m_token), &value));
+	logErrorIfNot(fromCString(m_token, (int)strlen(m_token), &value));
 	deserializeToken();
 }
